requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.72
botocore==1.35.72
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dill==0.3.6
dnslib==0.9.25
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
localstack==4.0.3
localstack-core==4.0.3
localstack-ext==4.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.2
plux==1.12.1
psutil==6.1.0
psycopg2==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyotp==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
semver==3.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tabulate==0.9.0
tailer==0.4.1
typer==0.14.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
